Bootloader: don't use debug message boxes on win32.
When running a DEBUG WINDOWED application on win32 we don't have any
console to send debug messages to. I find the repeating message boxes
annoying and propose to send the output text to the debugger.
If no debuggers are attached to the process DebugView tool
(https://docs.microsoft.com/en-us/sysinternals/downloads/debugview) can
be used.